separate commentContentSection from commentItem component
